+++ raytracerCUDA/sphere.cpp
@@ -1,14 +1,14 @@
 #include "headers/sphere.h"
 
-__device__ bool sphere::hit(const ray& r,double tmin,double tmax,intersection& isect) {
+__device__ bool sphere::hit(const ray& r,float tmin,float tmax,intersection& isect) {
 	vec3 OC = r.origin - center;
-	double a = dot(r.direction, r.direction);
-	double b = dot(OC, r.direction);
-	double c = dot(OC,OC)-radius*radius;
+	float a = dot(r.direction, r.direction);
+	float b = dot(OC, r.direction);
+	float c = dot(OC,OC)-radius*radius;
 
-	double disc = b * b - a * c;
+	float disc = b * b - a * c;
 	if (disc > 0) {
-		double temp = (-b-sqrt(disc)) / a;
+		float temp = (-b-sqrt(disc)) / a;
 		if (temp < tmax && temp > tmin) {
 			isect.hit_t = temp;
 			isect.hit_position = r.at(isect.hit_t);
